+++ SecretSauce/src/main/java/com/byoutline/secretsauce/SecretSauceSettings.kt
@@ -5,24 +5,29 @@ import android.content.Context
 import android.support.annotation.IdRes
 
 /**
- * Group of static variables that are required by SecretSauce classes.
+ * Initializes some variables globally for SecretSauce.
+ * As an alternative you usually can pass those values as method optional
+ * parameter.
  *
  * @author Sebastian Kacprzak <nait at naitbit.com>
  */
-object Settings {
+object SecretSauceSettings {
     var DEBUG: Boolean = false
-    var CONTAINER_VIEW_ID: Int = 0
-    var LOG_PREFIX = "SecretS"
+    var containerViewId: Int = 0
+    var brViewModelId: Int = 0
+    var logPrefix = "SecretS"
     var viewModelFactoryProvider: (ctx: Context) -> ViewModelProvider.Factory = {
         throw IllegalStateException("You must init viewModelFactoryProvider")
     }
 
 
     fun set(debug: Boolean = DEBUG,
-            @IdRes containerViewId: Int = CONTAINER_VIEW_ID,
+            @IdRes containerViewId: Int = this.containerViewId,
+            bindingViewModelId: Int = brViewModelId,
             viewModelFactoryProvider: (ctx: Context) -> ViewModelProvider.Factory = this.viewModelFactoryProvider) {
         DEBUG = debug
-        CONTAINER_VIEW_ID = containerViewId
+        this.containerViewId = containerViewId
+        brViewModelId = bindingViewModelId
         this.viewModelFactoryProvider = viewModelFactoryProvider
         System.setProperty("org.joda.time.DateTimeZone.Provider", "com.byoutline.secretsauce.utils.JdkBasedTimeZoneProvider")
     }