+++ CHANGELOG.md
@@ -11,6 +11,8 @@
      be done via `Settings.set(debug = BuildConfig.DEBUG, containerViewId = R.id.container)`
      * toast helpers are now `Context` extension functions
      * due to vast changes artifact name changed from `secretsauce` to `secretsaucekt`
+     * `Settings` renamed to `SecretSauceSettings` to make it more obvious in application what is being set.
+     * `ViewModelAutoLifecycle` and extensions functions added
   * 0.5.3 _06.05.2017_ Make listener interface in `CustomClickableSpan` public.
     `CustomFontTextView` now checks `suffixSpace` attribute that can be used to workaround
     wrong measurement on `wrap_content` when font is italic.